fix: implement security audit fixes for 18 of 31 findings
CRITICAL:
- GAPTOS-001: Fix NewEpochEvent serialization mismatch (serde_json → bcs)

HIGH:
- GAPTOS-002: Restore Noise handshake peer_id derivation check for untrusted peers
- GAPTOS-003: Restore on-chain RandomnessConfigSeqNum read (remove hardcoded 0)
- GAPTOS-004: Replace .unwrap() with ? in validator set deserialization
- GAPTOS-005: Return Result from DKG validator consensus info methods
- GAPTOS-006: Replace panic!() with Err for unknown JWK types

MEDIUM:
- GAPTOS-007: Replace todo!() with fallback RotatingProposer
- GAPTOS-010: Restore JWK sorting for non-gravity sources
- GAPTOS-011: Fix x25519 key generation to use provided RNG
- GAPTOS-012: Redact sensitive data from debug log
- GAPTOS-013: Add warning log for network address parse fallback

LOW:
- GAPTOS-015: Parse ChainId as u64 (not u8)
- GAPTOS-018: Handle GLOBAL_RELAYER.get() None gracefully
- GAPTOS-020: Replace eprintln!() with tracing::error!()
- GAPTOS-022: Implement Bcs::meta() (replace todo!())

INFO:
- GAPTOS-INFO-005: Replace impl Into with impl From
- GAPTOS-INFO-006: Rename error variant for BCS deserialization

+++ types/src/chain_id.rs
@@ -183,9 +183,9 @@ impl FromStr for ChainId {
     fn from_str(s: &str) -> Result<Self> {
         ensure!(!s.is_empty(), "Cannot create chain ID from empty string");
         NamedChain::str_to_chain_id(s).or_else(|_err| {
-            let value = s.parse::<u8>()?;
+            let value = s.parse::<u64>()?;
             ensure!(value > 0, "cannot have chain ID with 0");
-            Ok(ChainId::new(value as u64))
+            Ok(ChainId::new(value))
         })
     }
 }
